Ignore undocumented attribute writers
MSP-12718

`YARD::CLI::Stat` will mark writers created by `@!attribute` tags as
undocumented because the documentation under `@!attribute` is only
attached to the reader.  From the rendered documentation perspective,
the setter being undocumented doesn't matter because the reader and
writer are documented as one in the attribute documentation.  Therefore,
don't report writers for attributes as undocumented when `rake
yard:stats` is run.